core: change folder structures and fix exports (#10)
