+++ src/diffTime/index.ts
@@ -1,0 +1,8 @@
+export function getTimeDiff(date: string | Date, initDate?: string | Date): number{
+	date = new Date(date);
+	initDate = initDate ? new Date(initDate) : new Date();
+
+	const diff = date.getTime() - initDate.getTime();
+
+	return diff;
+}
